feat(murmurd): ignore erroneous exit when setting supw
Using -supw when invoking the murmur executable throws a fatal error and
exits the program. To bypass this, we use an OR operator and execute
`true`.

This issue is fixed in the upstream project by commit d8203ba94 [1];
this patch can be reverted in the future when a new version containing
the patch is released.

[1]: https://github.com/mumble-voip/mumble/commit/d8203ba94d528b092e0ff5a52a51af28f8f592f1

resolves sudoforge/docker-images#92
references mumble-voip/mumble#3911