feat: move mumble-server into the repository root
In commit c0cfda5340a2265c75900a5266bc4bdfd0fbcafc, this repository was
reorganized to support multiple independent projects. I've moved my
core development into a monorepo, and have made the decision that for
visibility and conciseness, this repository should be renamed to
`docker-mumble-server`. This decision was made primarily because this
repository has not been host to multiple images for quite some time, and
I have no intention changing that. Any additional images I develop in
the future will be published publicly in their own repository.