fix(mumble-server): use a workaround for the erroneous -supw failure
Using -supw currently throws a fatal error and exits the program. It has
been fixed in the upstream project by commit d8203ba94 [1], but has not
yet been included in a release. To bypass the issue and allow the entrypoint
script to proceed, we are ignoring the erroneous exit code by executing
`true` if the invocation fails.

This was previously fixed in commit b1229e09825, but was erroneously
removed in commit b7a0867ca93. Care should be taken to ensure the
upstream release _actually_ contains the fix before reverting this
commit.

[1]: https://github.com/mumble-voip/mumble/commit/d8203ba94d528b092e0ff5a52a51af28f8f592f1

closes #121
references #92 mumble-voip/mumble#3911 mumble-voip/mumble#3998 mumble-voip/mumble#4728